requirements.txt
llama-index==0.11.17
openai==1.51.2
pydantic==2.9.2
pyyaml==6.0.2
python-dotenv==1.0.1
vega-datasets==0.9.0
plotly==5.22.0
